feat: add new resolver for impacted files
