+++ graphql_api/tests/test_impacted_file.py
@@ -3,10 +3,12 @@ from dataclasses import dataclass
 from unittest.mock import PropertyMock, patch
 
 from django.test import TransactionTestCase
+from shared.reports.resources import Report, ReportFile, ReportLine
 from shared.torngit.exceptions import (
     TorngitClientGeneralError,
     TorngitObjectNotFoundError,
 )
+from shared.utils.sessions import Session
 
 from codecov_auth.tests.factories import OwnerFactory
 from compare.models import CommitComparison
@@ -31,24 +33,28 @@ query ImpactedFiles(
                             impactedFilesCount
                             indirectChangedFilesCount
                             impactedFiles {
-                                fileName
-                                headName
-                                baseName
-                                isNewFile
-                                isRenamedFile
-                                isDeletedFile
-                                isCriticalFile
-                                baseCoverage {
-                                    percentCovered
-                                }
-                                headCoverage {
-                                    percentCovered
-                                }
-                                patchCoverage {
-                                    percentCovered
-                                }
-                                changeCoverage
-                                    missesCount
+                                ... on ImpactedFiles {
+                                    results {
+                                        fileName
+                                        headName
+                                        baseName
+                                        isNewFile
+                                        isRenamedFile
+                                        isDeletedFile
+                                        isCriticalFile
+                                        baseCoverage {
+                                            percentCovered
+                                        }
+                                        headCoverage {
+                                            percentCovered
+                                        }
+                                        patchCoverage {
+                                            percentCovered
+                                        }
+                                        changeCoverage
+                                            missesCount
+                                        }
+                                    }
                                 }
                             }
                         }
@@ -212,6 +218,139 @@ class MockFileComparison(object):
         ]
 
 
+def sample_report():
+    report = Report(flags={"flag1": True})
+    first_file = ReportFile("foo/file1.py")
+    first_file.append(
+        1, ReportLine.create(coverage=1, sessions=[[0, 1]], complexity=(10, 2))
+    )
+    first_file.append(2, ReportLine.create(coverage=0, sessions=[[0, 1]]))
+    first_file.append(3, ReportLine.create(coverage=1, sessions=[[0, 1]]))
+    first_file.append(5, ReportLine.create(coverage=1, sessions=[[0, 1]]))
+    first_file.append(6, ReportLine.create(coverage=0, sessions=[[0, 1]]))
+    first_file.append(8, ReportLine.create(coverage=1, sessions=[[0, 1]]))
+    first_file.append(9, ReportLine.create(coverage=1, sessions=[[0, 1]]))
+    first_file.append(10, ReportLine.create(coverage=0, sessions=[[0, 1]]))
+    second_file = ReportFile("bar/file2.py")
+    second_file.append(12, ReportLine.create(coverage=1, sessions=[[0, 1]]))
+    second_file.append(
+        51, ReportLine.create(coverage="1/2", type="b", sessions=[[0, 1]])
+    )
+    third_file = ReportFile("file3.py")
+    third_file.append(1, ReportLine.create(coverage=1, sessions=[[0, 1]]))
+    report.append(first_file)
+    report.append(second_file)
+    report.append(third_file)
+    report.add_session(Session(flags=["flag1"]))
+    return report
+
+
+class TestImpactedFileFiltering(GraphQLTestHelper, TransactionTestCase):
+    def setUp(self):
+        self.org = OwnerFactory(username="codecov")
+        self.repo = RepositoryFactory(author=self.org, name="gazebo", private=False)
+        self.author = OwnerFactory()
+        self.parent_commit = CommitFactory(repository=self.repo)
+        self.commit = CommitFactory(
+            repository=self.repo,
+            totals={"c": "12", "diff": [0, 0, 0, 0, 0, "14"]},
+            parent_commit_id=self.parent_commit.commitid,
+        )
+        self.pull = PullFactory(
+            pullid=44,
+            repository=self.commit.repository,
+            head=self.commit.commitid,
+            base=self.parent_commit.commitid,
+            compared_to=self.parent_commit.commitid,
+        )
+        self.comparison = CommitComparisonFactory(
+            base_commit=self.parent_commit,
+            compare_commit=self.commit,
+            state=CommitComparison.CommitComparisonStates.PROCESSED,
+            report_storage_path="v4/test.json",
+        )
+        self.comparison_report = ComparisonReport(self.comparison)
+
+        self.query_impacted_files = """
+            query ImpactedFiles(
+                $org: String!
+                $repo: String!
+                $commit: String!
+                $filters: ImpactedFilesFilters
+            ) {
+                owner(username: $org) {
+                    repository(name: $repo) {
+                        ... on Repository {
+                            commit(id: $commit) {
+                                compareWithParent {
+                                    ... on Comparison {
+                                        impactedFiles(filters: $filters) {
+                                            ... on ImpactedFiles {
+                                                results {
+                                                    fileName
+                                                }
+                                            }
+                                            ... on UnknownFlags {
+                                                message
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        """
+
+    @patch("services.report.build_report_from_commit")
+    @patch("services.archive.ArchiveService.read_file")
+    @patch("services.comparison.Comparison.git_comparison")
+    def test_filtering_with_successful_flags(
+        self, git_comparison_mock, read_file, build_report_from_commit
+    ):
+        git_comparison_mock.return_value = None
+        build_report_from_commit.return_value = sample_report()
+        read_file.return_value = mock_data_from_archive
+
+        variables = {
+            "org": self.org.username,
+            "repo": self.repo.name,
+            "commit": self.commit.commitid,
+            "filters": {"flags": ["flag1"]},
+        }
+
+        data = self.gql_request(self.query_impacted_files, variables=variables)
+        assert (
+            "results"
+            in data["owner"]["repository"]["commit"]["compareWithParent"][
+                "impactedFiles"
+            ]
+        )
+
+    @patch("services.report.build_report_from_commit")
+    @patch("services.archive.ArchiveService.read_file")
+    @patch("services.comparison.Comparison.git_comparison")
+    def test_filtering_with_unknown_flags(
+        self, git_comparison_mock, read_file, build_report_from_commit
+    ):
+        git_comparison_mock.return_value = None
+        build_report_from_commit.return_value = sample_report()
+        read_file.return_value = mock_data_from_archive
+
+        variables = {
+            "org": self.org.username,
+            "repo": self.repo.name,
+            "commit": self.commit.commitid,
+            "filters": {"flags": ["fake_flag"]},
+        }
+
+        data = self.gql_request(self.query_impacted_files, variables=variables)
+        assert data["owner"]["repository"]["commit"]["compareWithParent"][
+            "impactedFiles"
+        ] == {"message": "No coverage with chosen flags"}
+
+
 class TestImpactedFile(GraphQLTestHelper, TransactionTestCase):
     def setUp(self):
         self.org = OwnerFactory(username="codecov")
@@ -268,46 +407,48 @@ class TestImpactedFile(GraphQLTestHelper, TransactionTestCase):
                         "compareWithParent": {
                             "impactedFilesCount": 2,
                             "indirectChangedFilesCount": 1,
-                            "impactedFiles": [
-                                {
-                                    "fileName": "fileA",
-                                    "headName": "fileA",
-                                    "baseName": "fileA",
-                                    "isNewFile": False,
-                                    "isRenamedFile": False,
-                                    "isDeletedFile": False,
-                                    "isCriticalFile": False,
-                                    "baseCoverage": {
-                                        "percentCovered": 41.666666666666664
+                            "impactedFiles": {
+                                "results": [
+                                    {
+                                        "fileName": "fileA",
+                                        "headName": "fileA",
+                                        "baseName": "fileA",
+                                        "isNewFile": False,
+                                        "isRenamedFile": False,
+                                        "isDeletedFile": False,
+                                        "isCriticalFile": False,
+                                        "baseCoverage": {
+                                            "percentCovered": 41.666666666666664
+                                        },
+                                        "headCoverage": {
+                                            "percentCovered": 85.71428571428571
+                                        },
+                                        "patchCoverage": {"percentCovered": 50.0},
+                                        "changeCoverage": 44.047619047619044,
+                                        "missesCount": 1,
                                     },
-                                    "headCoverage": {
-                                        "percentCovered": 85.71428571428571
+                                    {
+                                        "fileName": "fileB",
+                                        "headName": "fileB",
+                                        "baseName": "fileB",
+                                        "isNewFile": False,
+                                        "isRenamedFile": False,
+                                        "isDeletedFile": False,
+                                        "isCriticalFile": False,
+                                        "baseCoverage": {
+                                            "percentCovered": 41.666666666666664
+                                        },
+                                        "headCoverage": {
+                                            "percentCovered": 85.71428571428571
+                                        },
+                                        "patchCoverage": {
+                                            "percentCovered": 85.71428571428571
+                                        },
+                                        "changeCoverage": 44.047619047619044,
+                                        "missesCount": 2,
                                     },
-                                    "patchCoverage": {"percentCovered": 50.0},
-                                    "changeCoverage": 44.047619047619044,
-                                    "missesCount": 1,
-                                },
-                                {
-                                    "fileName": "fileB",
-                                    "headName": "fileB",
-                                    "baseName": "fileB",
-                                    "isNewFile": False,
-                                    "isRenamedFile": False,
-                                    "isDeletedFile": False,
-                                    "isCriticalFile": False,
-                                    "baseCoverage": {
-                                        "percentCovered": 41.666666666666664
-                                    },
-                                    "headCoverage": {
-                                        "percentCovered": 85.71428571428571
-                                    },
-                                    "patchCoverage": {
-                                        "percentCovered": 85.71428571428571
-                                    },
-                                    "changeCoverage": 44.047619047619044,
-                                    "missesCount": 2,
-                                },
-                            ],
+                                ]
+                            },
                         }
                     }
                 }
