+++ graphql_api/types/comparison/comparison.graphql
@@ -1,7 +1,7 @@
 type Comparison {
   state: String!
   impactedFile(path: String!): ImpactedFile
-  impactedFiles(filters: ImpactedFilesFilters): [ImpactedFile]!
+  impactedFiles(filters: ImpactedFilesFilters): ImpactedFilesResult!
   impactedFilesDeprecated(filters: ImpactedFilesFilters): [ImpactedFile]!
   impactedFilesCount: Int!
   indirectChangedFilesCount: Int!
@@ -15,7 +15,6 @@ type Comparison {
   hasDifferentNumberOfHeadAndBaseReports: Boolean!
   flagComparisonsCount: Int!
   componentComparisonsCount: Int!
-  impactedFilesNew(filters: ImpactedFilesFilters): ImpactedFilesResult!
 }
 
 type MissingBaseCommit implements ResolverError {
